refact: use Robolectric at :platform:autos-test
