+++ feature/composer/build.gradle.kts
@@ -33,6 +33,5 @@ dependencies {
   implementation(libs.android.compose.material.icons)
 
   testImplementation(project(":platform:autos-test"))
-  testImplementation(libs.android.compose.ui.test.junit)
   testImplementation(libs.robolectric)
 }
