+++ platform/autos-test/build.gradle.kts
@@ -33,10 +33,11 @@ dependencies {
   androidTestImplementation(libs.android.compose.ui.test.manifest)
   androidTestImplementation(libs.kotlin.test)
 
+  api(libs.android.compose.ui.test.junit)
+  api(libs.android.test.espresso.core)
+
   implementation(project(":ext:reflection"))
   implementation(project(":platform:autos"))
-  implementation(libs.android.compose.ui.test.junit)
-  implementation(libs.android.test.espresso.core)
 
   testImplementation(project(":platform:testing"))
   testImplementation(libs.android.compose.ui.test.manifest)
