+++ platform/autos-test/src/test/java/br/com/orcinus/orca/platform/autos/test/kit/input/text/SemanticsNodeInteractionsProviderExtensionsTests.kt
@@ -24,7 +24,10 @@ import br.com.orcinus.orca.platform.autos.kit.input.text.error.rememberErrorDisp
 import br.com.orcinus.orca.platform.autos.theme.AutosTheme
 import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
 
+@RunWith(RobolectricTestRunner::class)
 internal class SemanticsNodeInteractionsProviderExtensionsTests {
   @get:Rule val composeRule = createComposeRule()
